+++ app/src/main/java/com/example/appfrisaahorasi/pantallas/Perfil/mapScreen.kt
@@ -1,0 +1,38 @@
+package com.example.appfrisaahorasi.pantallas.Perfil
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.viewinterop.AndroidView
+import androidx.navigation.NavController
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+
+@Composable
+fun MapScreen(navController: NavController) {
+    val mapView = rememberMapViewWithLifecycle()
+
+    // on below line creating a
+    // column for our maps.
+    Column(
+        modifier = Modifier
+            .fillMaxHeight()
+            .fillMaxWidth()
+            .background(Color.White)
+    ) {
+        // on below line adding a map view to it.
+        AndroidView({ mapView }) { mapView ->
+            // on below line launching our map view
+            CoroutineScope(Dispatchers.Main).launch {
+                //val map = mapView.awaitMap()
+                // on below line adding zoom controls for map.
+                //map.uiSettings.isZoomControlsEnabled = true
+            }
+        }
+    }
+}
